requirements.txt
# Core Dependencies
flask
flask-cors
pymongo
chromadb
langchain
langchain-openai
langchain-google-genai
langgraph
nicegui
httpx
langchain-chroma
google-generativeai
pydantic
requests


# Observability Stack
sentry-sdk[flask]
langfuse
prometheus-flask-exporter
prometheus-client
python-dotenv
